VietNameOCR/requirements.txt
torch
torchvision
pyyaml
Pillow
numpy
tqdm